Add initial copy of OpAMP spec (#2)
The README.md file is originally Apache 2.0-Licensed. It is an exact copy of
https://github.com/signalfx/opamp-spec/blob/main/README.md.

According to Apache License 2.0 requirements and OpenTelemetry guidelines
(https://github.com/open-telemetry/community/blob/main/CONTRIBUTING.md) we
are going to retain the original copyright notice.

As recommended by the guidelines we will add "Copyright The OpenTelemetry Authors"
notice once we make any changes to this file.