Add version field to Addons
Contributes to https://github.com/open-telemetry/opamp-spec/issues/31

The Addon version is added to the Server offers and to the Agent Status reports.

This covers the use case described in the issue:
"Agent downloads a different addon hash/version, the addon errors, the agent is
 able to use the previous addon hash/version".

This may or may not cover the second use case:
"Agent uses multiple addon hashes/versions concurrently, using capabilities that
 may have been removed and added across addon hashes/version".
The protocol does not allow reporting status of multiple versions concurrently,
but the Agent can still keep multiple versions around. It is not completely
clear from the description of the use case if this is sufficient.