Generalize Agent identification (#48)
Resolves https://github.com/open-telemetry/opamp-spec/issues/16

This change allows the protocol to be used with non-standalone
Agents, such as OpenTelemetry SDK. For such Agents the identification
is not done via `service.*` attributes, but via a different set of
attributes (e.g. `telemetry.sdk.*` attributes).

This change replaces (agent_type,agent_version) tuple by a list of
attributes which can be freely chosen by the Agent. The list is
further split into 2 parts: identifying, which is required for linking
to own telemetry, and non-identifying which allow additional description.

We also need to submit a PR against Otel spec repo to add uniquely
identifying attributes for Otel SDKs (e.g. telemetry.sdk.instance.id).