Separate the notions of Agent and Client (#122)
Separating the previous notion of Agent into Agent (as a managed entity) and Client (as an endpoint of OpAMP).

We have heard from a few people that it was confusing and unclear how the Agent and the Client are related. This makes clarifications and makes the Client and the Agent more distinct in the spec.