Move proto files to opamp-spec repo (#126)
Contributes to https://github.com/open-telemetry/opamp-go/issues/133

Now that we are adding OpAMP Java implementation the protos can no longer
live in opamp-go. We are moving the protos to the spec repo since it fits
this repo's purpose naturally.

The proto files are copied verbatim from opamp-go. Any changes needed in
the proto files can be done in a future PR.

In the future it is also worth exploring if we can delete the message definitions
from specification.md and instead link them directly to the proto files so
that we have a single source of truth.