Rewrite security recommendation to be non-normative
Usage of format requirement level implies somehow the recommendations
are important part of the protocol. I think it is misleading. These
are really recommendations, and certainly they are not complete or
sufficient.

I think we should de-emphasize them a bit and really just keep
as what they are: recommendations and warnings.