Replace ULIDs by 16 byte ids and recommend UUID v7 (#186)
The early version of the spec was written before UUID v7 existed and
ULID was chosen as a ID type that is suited to be used as a primary
key in databases. Since than UUID v7 was standardized and aims to
serve the exact same niche.

## Proposal

Replace instance_id definition to be an arbitrary sequence of 16 bytes.
Both ULID and any version of UUID may be used as the value. We also
recommend to use UUID v7 for the value.

This is a breaking change in the spec. We recommend opamp-go implementation
to implement the change in backward compatible way, supporting both old and
new instance_uid formats for a period of time.

Since `bytes` and `string` are compatible in the Protobuf wire encoding
it is possible to use the new Protobuf declarations to receive messages
encoded using the old Protobuf declarations - string's UTF8 bytes
will simply populate the bytes of the new field. Receivers can simply
probe for the length of the received bytes field and if it is 26 treat
it as ULID in canonical format, if it is 16 bytes treat it as opaque
id in new format, otherwise consider it invalid.

I will also post a PR in opamp-go that implements the above behavior.